Move alternation into its own file.
